refactor: Step 2 - Introduce foundational Rust modules
- Extract logging initialisation into `logging.rs` and update `lib.rs` to call the helper.
- Move session lifecycle logic from `lib.rs` into a new `session.rs`, keeping function signatures untouched and re-exporting via `lib.rs`.
- Update module declarations and adjust imports; verify tests.
