+++ .gitignore
@@ -7,3 +7,5 @@ build
 *~
 .idea/
 .cargo/
+
+**/*.egg-info/
