+++ design-docs/file-level-srp-refactor-plan.status.md
@@ -1,0 +1,8 @@
+# File-Level SRP Refactor Status
+
+## Current Status
+- ✅ Step 2 complete: introduced `src/logging.rs` for one-time logger initialisation and migrated tracing session lifecycle (`start_tracing`, `stop_tracing`, `is_tracing`, `flush_tracing`, `ACTIVE` flag) into `src/session.rs`, with `src/lib.rs` now limited to PyO3 wiring and re-exports.
+- ⚠️ Test baseline still pending: `cargo check` succeeds; `cargo test` currently fails to link in the sandbox because CPython development symbols are unavailable, matching the pre-refactor limitation.
+
+## Next Task
+- Step 3: Restructure runtime tracer internals by creating `src/runtime/mod.rs` and extracting activation control, value encoding, and writer/output-path handling into focused submodules before reconnecting them through the new façade.
